[features] write tests, some update and fix

+++ src/Controller/TaskController.php
@@ -4,12 +4,21 @@ namespace App\Controller;
 
 use App\Entity\Task;
 use App\Form\TaskType;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
+use Symfony\Component\Security\Core\Security;
 
-class TaskController extends Controller
+class TaskController extends AbstractController
 {
+    private $security;
+
+    public function __construct(Security $security)
+    {
+        $this->security = $security;
+    }
+
     /**
      * @Route("/tasks", name="task_list")
      */
@@ -83,6 +92,10 @@ class TaskController extends Controller
      */
     public function deleteTaskAction(Task $task)
     {
+        if (null === $task->getAuthor() && false === $this->security->isGranted('ROLE_ADMIN') || $task->getAuthor() !== $this->getUser()) {
+            throw new AccessDeniedHttpException("Vous n'avez pas l'autorisation pour supprimer cette tâche");
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($task);
         $em->flush();
